First stage DONE. UserService, OpenAIClient, plainTextParser, LamgChain works
