+++ packages/llm/clients/open-ai/index.ts
@@ -1,0 +1,2 @@
+export { default as OpenAIClient } from "./openai.client";
+export * from "./openai.interface";
